Move panel commands to their own file.
